feat:반응형 구현
